Add EWMA calculation method for training load
First implementation of EWMA (exponentially weighted moving average). It
weighs more recent training loads heavier. Training session 3 weeks ago
isn't affecting fatigue as much as session 3 days ago.

+++ backend/TrackerWebAPI/Models/LoadingStatusSnapshotDTO.cs
@@ -8,7 +8,7 @@
 
     public class LoadingStatusSnapshotDTO
     {
-        public LoadingStatusSnapshotDTO(int acute, int chronic, float ratio, WorkloadCalculateMethod method, DateTime snapshotDate, int dailyLoad)
+        public LoadingStatusSnapshotDTO(double acute, double chronic, double ratio, WorkloadCalculateMethod method, DateTime snapshotDate, int dailyLoad)
         {
             Method = method;
             Acute = acute;
@@ -19,9 +19,9 @@
         }
         public WorkloadCalculateMethod Method { get; set; }
         public DateTime SnapshotDate { get; set; }
-        public int Acute { get; set; }
-        public int Chronic { get; set; }
-        public float Ratio { get; set; }
+        public double Acute { get; set; }
+        public double Chronic { get; set; }
+        public double Ratio { get; set; }
         public int DailyLoad { get; set; }
     }
 }
